% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranking_seqs.R
\name{combine_ranking_seqs}
\alias{combine_ranking_seqs}
\title{combine_ranking_seqs}
\usage{
combine_ranking_seqs(rank_seqs, method = "quantile", q = 0.25)
}
\arguments{
\item{rank_seqs}{A ranking sequence of all the variables}

\item{method}{The ensemble method "median"/"arith.mean"/"geom.mean"/"quantile"}

\item{q}{Required only when method is "quantile", specify the quantile value}
}
\value{
A vector (length is number of potential shift postions). The ensembled ranking sequence
}
\description{
ensemble the ranking sequnces with median/arith.mean/geom.mean/quantile
}
