% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELPASO.R
\name{forward_backward_selection}
\alias{forward_backward_selection}
\title{forward_backward_selection}
\usage{
forward_backward_selection(
  tree,
  Y,
  seq,
  increasing = TRUE,
  criterions,
  stops,
  maxShifts,
  direction = c("same", "opposite"),
  opt
)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder}

\item{Y}{Y trait vector without missing entries}

\item{seq}{A vector gives that selection order of the shifts. (For example, the ensembled rank)}

\item{increasing}{If a smaller value in param seq means more important variable, increasing = TRUE, vice versa.}

\item{criterions}{A two-element vector. The model selection criterions for forward selection and backward selection respectively}

\item{stops}{A two-element vector. Each element is TRUE/FALSE. If the selection process will be stopped if score stops decreasing}

\item{maxShifts}{the maximum number of shifts}

\item{direction}{'same'(default)/'opposite'. If the forward selection and the backward selection are produced in the same direction.}

\item{opt}{options for l1ou functions}
}
\value{
\item{s.c}{The position of shifts}
\item{score}{The criterion score}
}
\description{
This function provides forward+bacward selection given a variable ranking sequence
}
