% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELPASO.R
\name{forward_selection}
\alias{forward_selection}
\title{forward_selection}
\usage{
forward_selection(
  tree,
  Y,
  seq,
  increasing = TRUE,
  criterion = c("BIC", "pBIC", "pBICess", "AICc", "mBIC"),
  stop = TRUE,
  maxShifts = 20,
  opt
)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder}

\item{Y}{Y trait vector without missing entries}

\item{seq}{A vector gives that selection order of the shifts. (For example, the ensembled rank)}

\item{increasing}{If a smaller value in param seq means more important variable, increasing = TRUE, vice versa.}

\item{criterion}{the model selection criterion}

\item{stop}{TRUE/FALSE. If the selection process will be stopped if score stops decreasing}

\item{maxShifts}{the maximum number of shifts}

\item{opt}{options for l1ou functions}
}
\value{
\item{s.c}{The position of shifts}
\item{score}{The criterion score}
}
\description{
This function provides forward selection given a variable ranking sequence
}
