% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_likelihood.R
\name{get_prediction_likelihood}
\alias{get_prediction_likelihood}
\title{get_prediction_likelihood}
\usage{
get_prediction_likelihood(
  tree,
  Y,
  selected_variable,
  alpha,
  test_data,
  xtype = c("simpX", "orgX")
)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder}

\item{Y}{Y trait vector without missing entries}

\item{selected_variable}{The variables selected}

\item{alpha}{The selective rate}

\item{test_data}{A bunch of testing data given by the true model}

\item{xtype}{the type of design matrix}
}
\value{
The prediction log likelihood value
}
\description{
This function calculates the prediction log likelihood to evaluate the accuracy of estimates
}
