% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranking_seqs.R
\name{get_ranking_seqs}
\alias{get_ranking_seqs}
\title{get_ranking_seqs}
\usage{
get_ranking_seqs(
  tree,
  Y,
  nsamples,
  size,
  replace = FALSE,
  alpha,
  sigma2,
  sigma2_error = 0,
  xtype = c("simpX", "orgX"),
  penalty = c("LASSO", "SCAD")
)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder}

\item{Y}{Y trait vector without missing entries}

\item{nsamples}{The number subsamples are generated. Integer.}

\item{size}{The size of each subsample. Integer.}

\item{replace}{if replace while subsampling, TRUE/FALSE(default)}

\item{alpha}{the adaption rate (known or estimated)}

\item{sigma2}{the variance of OU model (known or estimated)}

\item{sigma2_error}{the variance of gaussian measurement error}

\item{xtype}{type of design matrix, 'simpX'(simple design matrix, default) or 'orgX'(original design matrix)}

\item{penalty}{Penalty function to be applied. Either "LASSO"(default) or "SCAD"}
}
\value{
A matrix, each row is a ranking sequence of variables, each column represents the rank sets of each variable
}
\description{
This function generates subsamples of a given combination of tree and trait values, applys LASSO/SCAD on each subsample, and produces ranking sequence of each subsample
}
