% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_likelihood.R
\name{get_test_data}
\alias{get_test_data}
\title{get_test_data}
\usage{
get_test_data(
  tree,
  true_variable,
  beta,
  alpha,
  n_test,
  xtype = c("simpX", "orgX")
)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder}

\item{true_variable}{The true shifts}

\item{beta}{The coefficients of shifts}

\item{alpha}{The selective rate}

\item{n_test}{The number of testing datasets}

\item{xtype}{the type of design matrix}
}
\value{
A matrix, each row is a testing dataset for the true model
}
\description{
This function generates a bunch of testing data given the tree and true shifts
}
